requirements.txt
backcall==0.1.0
cycler==0.10.0
decorator==4.4.0
future==0.18.2
gitdb2==2.0.5
GitPython==2.1.11
h5py==2.10.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.15.1
kiwisolver==1.1.0
matplotlib==3.0.2
numpy==1.16.3
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.3.0
progress==1.5
prompt-toolkit==2.0.10
pthflops==0.2.2
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.23.1
scipy==1.4.1
six==1.12.0
sklearn==0.0
smmap2==2.0.5
threadpoolctl==2.1.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.46.1
traitlets==4.3.3
wcwidth==0.1.7
